Add `CyclotomicField` to represent cyclotomic operations

+++ ff/src/biginteger/arithmetic.rs
@@ -52,7 +52,7 @@ pub(crate) fn mac_discard(a: u64, b: u64, c: u64, carry: &mut u64) {
     *carry = (tmp >> 64) as u64;
 }
 
-pub fn find_wnaf(num: &[u64]) -> Vec<i64> {
+pub fn find_wnaf(num: &[u64]) -> Vec<i8> {
     let is_zero = |num: &[u64]| num.iter().all(|x| *x == 0u64);
     let is_odd = |num: &[u64]| num[0] & 1 == 1;
     let sub_noborrow = |num: &mut [u64], z: u64| {
@@ -87,9 +87,9 @@ pub fn find_wnaf(num: &[u64]) -> Vec<i64> {
     let mut res = vec![];
 
     while !is_zero(&num) {
-        let z: i64;
+        let z: i8;
         if is_odd(&num) {
-            z = 2 - (num[0] % 4) as i64;
+            z = 2 - ((num[0] % 4) as i8);
             if z >= 0 {
                 sub_noborrow(&mut num, z as u64)
             } else {
